+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/DataStreamer.cs
@@ -57,5 +57,7 @@ public partial class DataStreamer : Form, IDemoWindow
                 formsPlot1.Interaction.Enable();
             }
         };
+
+        cbContinuouslyAutoscale.CheckedChanged += (s, e) => Streamer.ContinuouslyAutoscale = cbContinuouslyAutoscale.Checked;
     }
 }