+++ CHANGELOG.md
@@ -7,6 +7,7 @@ _Not yet on NuGet..._
 * Scatter: Disable marker outline visibility by default (#3720)
 * Markers: Disable rendering of lines when `LineWidth` is `0` (#3720)
 * Scatter: Added support for filling above and below the curve (#3318, #3380) @xichaoqiang @Diddlik @slotinvo
+* DataStreamer: Added `ContinuouslyAutoscale` flag to allow the vertical range to always tightly fit the data (#3561) @hazenjaqdx3 @zhhding @Xhichn
 
 ## ScottPlot 5.0.27
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-04-23_